requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
scipy>=1.7.0

# Chemical informatics
rdkit>=2022.3.0
padelpy>=0.1.14

# Machine learning
xgboost>=1.5.0
lightgbm>=3.3.0

# Configuration and data handling
PyYAML>=6.0
joblib>=1.1.0

# Visualization
plotly>=5.0.0
jupyter>=1.0.0

# Development and testing
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Optional: for advanced features
torch>=1.10.0  # For deep learning models
tensorflow>=2.8.0  # Alternative deep learning framework
umap-learn>=0.5.0  # For dimensionality reduction
shap>=0.41.0  # For model interpretability